Make controlsfor player a bit more robust
Adds the CharacterController component to the Player prefab, and uses that to move
the character based on the direction the player is moving it. Adds a FreeLookMovementSpeed
variable that can be adjusted in the Player prefab object from the editor window. Adds
logic to the player object that will face the character in the direction it is moving.