Add idle / run aniamtions to character movement
Adds a Blend tree to the animator associated with the Player prefab which currently
only contains two animations, one for idle and one for running. Add in code to the
PlayerTestScript where the movement logic is currently held that will transition between
the two animations based on whether the player is pressing a movement key or not.